Fixed a compilation error

+++ include/lvox/scanner/scan.hpp
@@ -11,10 +11,10 @@ class Beam;
 class Scan
 {
   public:
-    virtual ~Scan()                                                       = default;
-    virtual auto get_points() const -> const PointCloudView&              = 0;
-    virtual auto get_beams() const -> std::vector<lvox::Beam>             = 0;
-    virtual auto get_scan_position(size_t gps_time) const -> const Point& = 0;
+    virtual ~Scan()                                                     = default;
+    virtual auto get_points() const -> const PointCloudView&            = 0;
+    virtual auto get_beams() const -> std::vector<lvox::Beam>           = 0;
+    virtual auto get_scan_position(long gps_time) const -> const Point& = 0;
 };
 
 } // namespace lvox
